+++ src/app/admin/Dashboard/page.tsx
@@ -7,25 +7,27 @@ import Sidebar from "../components/Sidebar";
 
 const Dashboard = () => {
   return (
-    <>
+    <div className="flex h-screen">
       <Sidebar />
-      <Topbar
-        title={"Dashboard"}
-        subtitle={"Total Overview"}
-        description={`Track key metrics - A real-time snapshot of your platform's performance.`}
-      />
-      <div className="p-6">
-        <OverviewCards />
-        <div className="grid grid-cols-1 lg:grid-cols-7 gap-6">
-          <div className="lg:col-span-5">
-            <WaitlistTable />
-          </div>
-          <div className="lg:col-span-2">
-            <LatestBlogs />
+      <div className="h-screen overflow-y-auto">
+        <Topbar
+          title={"Dashboard"}
+          subtitle={"Total Overview"}
+          description={`Track key metrics - A real-time snapshot of your platform's performance.`}
+        />
+        <div className="p-6">
+          <OverviewCards />
+          <div className="grid grid-cols-1 lg:grid-cols-7 gap-6">
+            <div className="lg:col-span-5">
+              <WaitlistTable />
+            </div>
+            <div className="lg:col-span-2">
+              <LatestBlogs />
+            </div>
           </div>
         </div>
       </div>
-    </>
+    </div>
   );
 };
