+++ src/app/admin/components/Sidebar.tsx
@@ -47,7 +47,7 @@ export default function AppSidebar() {
             </MenuItem>
           </Link>
 
-          <Link href="/admin/mail">
+          <Link href="/admin/mailManagement">
             <MenuItem
               icon={<FaEnvelope />}
               className="text-white bg-black/60 rounded-md mx-auto border-white/30 border hover:bg-white/90 hover:text-black/80 my-1"
