+++ src/app/admin/components/newletter/NewsletterPage.tsx
@@ -58,7 +58,7 @@ const NewsletterPage = () => {
   };
 
   return (
-    <div className="pe-4 flex flex-col gap-3 bg-white/40 min-h-screen">
+    <div className="pe-4 flex-1 flex flex-col  bg-white/40 h-screen overflow-y-auto">
       <Topbar
         title="Newsletters"
         subtitle="Newsletters"
