+++ src/app/admin/mailManagement/page.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import AppSidebar from "../components/Sidebar";
+import MailManagementPage from "../components/mailcomp/MailManagementPage";
+
+const MailHome = () => {
+  return (
+    <div className="flex">
+      <AppSidebar />
+      <MailManagementPage />
+    </div>
+  );
+};
+
+export default MailHome;
